requirements.txt
# requirements.txt
streamlit
google-api-python-client
boto3
python-dotenv
